requirements.txt
Flask==1.1.2
Flask-Cors==3.0.9
Jinja2==2.11.2
MarkupSafe==1.1.1
Werkzeug==1.0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
idna==2.10
itsdangerous==1.1.0
newsapi==0.1.1
numpy==1.19.2
pandas==1.1.2
pip==20.2.3
python-dateutil	==2.8.1
pytz==2020.1
requests==2.24.0
setuptools==50.3.0
six==1.15.0
tiingo==0.12.0
urllib3==1.25.10